Accepting merge request

+++ src/main/java/edu/matc/pricecheck/Grocery.java
@@ -2,24 +2,25 @@ package edu.matc.pricecheck;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import javax.annotation.Generated;
+import java.math.BigDecimal;
 
 @Generated("com.robohorse.robopojogenerator")
 public class Grocery{
 
 	@JsonProperty("latitude")
-	private double latitude;
+	private BigDecimal latitude;
 
 	@JsonProperty("name")
 	private String name;
 
 	@JsonProperty("longtitude")
-	private double longtitude;
+	private BigDecimal longtitude;
 
-	public void setLatitude(double latitude){
+	public void setLatitude(BigDecimal latitude){
 		this.latitude = latitude;
 	}
 
-	public double getLatitude(){
+	public BigDecimal getLatitude(){
 		return latitude;
 	}
 
@@ -31,11 +32,11 @@ public class Grocery{
 		return name;
 	}
 
-	public void setLongtitude(double longtitude){
+	public void setLongtitude(BigDecimal longtitude){
 		this.longtitude = longtitude;
 	}
 
-	public double getLongtitude(){
+	public BigDecimal getLongtitude(){
 		return longtitude;
 	}
